backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.3
h11==0.16.0
idna==3.11
motor==3.7.1 
passlib[bcrypt]==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.4
python-dotenv==1.2.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
openai